+++ Online Shopping System/src/main/java/com/tothenew/entities/product/ProductReviewKey.java
@@ -8,6 +8,6 @@ import java.io.Serializable;
 @Embeddable
 @Data
 public class ProductReviewKey implements Serializable {
-    private int productId;
-    private int customerId;
+    private Long productId;
+    private Long customerId;
 }
